rename tests.util to tests.utils
